Ignore missing event handler.
There might be events in an aggregate stream that do not (yet)
controbite to sourced state and therefore might not have apply_ handler
implementation. Adding empty methods to dodge this behaviour is
tedious.

One can add own strategy to deal with it in simillar way. I believe
that is just better default.